singup & login page setup
