+++ components/success/page.tsx
@@ -1,27 +1,26 @@
-'use client'
-import React from 'react'
-import Lottie from "lottie-react";
-import GifFile from '@/constants/success.json'
+import { successPageContent } from "@/constants/formContent";
+import Banner from "../Banner/page";
+import FormLayout from "../FormLayout/page";
+import Logo from "../Logo/page";
+import GifComponent from "./GifComponent";
+import loginBanner from "../../public/assets/loginBanner.png";
+
 
 const page = () => {
     return (
-        <div className='flex flex-col items-center justify-center' >
-           <div className='w-[60%] h-[60%]'>
-           <Lottie
-        loop={true}
-        autoplay={true}
-        animationData={GifFile}
-        height={50}
-        width={50}
-      />         
-            </div> 
-            <p className='text-base font-medium mt-3 leading-5 text-center text-black'>We`ll handle the rest with our technolagy,and one of our team member contact you shortly.</p>
+        <main className="px-16 pb-24 bg-[#F1F1F1] max-md:px-5 justify-center items-center flex flex-col font-libre">
+            <div className="flex w-full pt-4 items-start px-4">
+                <Logo />
+            </div>
 
-            <button  className="py-2.5 mt-8 max-w-full text-sm md:text-base font-medium items-center text-center text-black  bg-amber-300 rounded-3xl w-[216px] max-md:px-5">
-                <a href='/' className=''>BACK TO HOMEPAGE</a>
-            </button>
-        </div>
-    );
+            <div className="flex lg:gap-16 md:gap-0 sm:gap-0 max-md:flex-col w-[95%]">
+                <Banner content={loginBanner.src} />
+                <div></div>
+                <FormLayout content={successPageContent} component={<GifComponent />} />
+            </div>
+        </main>
+    )
+
 };
